Moved the code to run_store.go Now the call chain is run_client->run_server->resource_manager->run_store
