+++ backend/src/apiserver/resource/resource_manager.go
@@ -264,6 +264,10 @@ func (r *ResourceManager) ListJobs(filterContext *common.FilterContext,
 	return r.jobStore.ListJobs(filterContext, opts)
 }
 
+func (r *ResourceManager) TerminateRun(runId string) error {
+	return r.runStore.TerminateRun(runId)
+}
+
 func (r *ResourceManager) GetJob(id string) (*model.Job, error) {
 	return r.jobStore.GetJob(id)
 }
